% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_managment.R
\name{activeteam2dcf}
\alias{activeteam2dcf}
\title{Create slackr dcf file}
\usage{
activeteam2dcf(file = "~/.slackr", verbose = TRUE)
}
\arguments{
\item{file}{character, path to write the dcf file to, Default: '~/.slackr'}

\item{verbose}{logical, Print messages to console, Default: TRUE}
}
\description{
Convert the active team to a slackr compatible dcf file on the local system
}
\details{
If the file is "" then the output will be printed to the console
}
\concept{files}
