% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_management.R
\name{add_team_code}
\alias{add_team_code}
\title{Semi-Interactive Team Management}
\usage{
add_team_code(code, redirect_uri = NULL, verbose = TRUE)
}
\arguments{
\item{code}{character, a code returned by the slack oauth2 v2 api.}

\item{redirect_uri}{character, the uri to which the user was redirected when
the code was generated.}

\item{verbose}{logical, Print messages to console, Default: TRUE}
}
\value{
The token (invisibly)
}
\description{
Add a team using a Slack response code.
}
\details{
Launch a browser window to interactively grant slackteams permission
to act on your behalf on a Slack team.

Two environment variables control this function:
\itemize{
\item{SLACK_CLIENT_ID: character, the client_id of a Slack app. If this
is not provided, the function will use the built-in R4DS Slack app.}
\item{SLACK_CLIENT_SECRET: character, the client_secret of a Slack app.
If this is not provided, the function will use the built-in R4DS Slack
app.}
}
}
\note{
This function does not currently work in an Rstudio Server setup. We
are exploring options to remedy this situation.
}
\concept{management}
