% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_managment.R
\name{add_team_interactive}
\alias{add_team_interactive}
\title{Interactive Team Managment}
\usage{
add_team_interactive(scopes = load_scopes())
}
\arguments{
\item{scopes}{character, scopes to request. Must include "users:read",
"channels:read", "groups:read", "im:read", and "mpim:read" at minimum.}
}
\description{
Add a team interactively
}
\details{
Launch a browser window to interactively grant slackteams permission
to act on your behalf on a Slack team.

Two global package options control this function:
\itemize{
\item{slack_client_id: character, the client_id of a Slack app. If this
is not provided, the function will use the built-in R4DS Slack app.}
\item{slack_client_secret: character, the client_secret of a Slack app.
If this is not provided, the function will use the built-in R4DS Slack
app.}
}
}
\note{
This function does not currently work in an Rstudio Server setup. We
are exploring options to remedy this situation.
}
\concept{management}
