% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_managment.R
\name{add_team_token}
\alias{add_team_token}
\title{Add team credentials}
\usage{
add_team_token(team, token, scopes)
}
\arguments{
\item{team}{character, team name}

\item{token}{character, api token issued by slack}

\item{scopes}{token scopes}
}
\description{
Add a team with a user specified token
and the scopes for it.
}
