% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_managment.R
\name{auth_url}
\alias{auth_url}
\title{Slack Authorization URL}
\usage{
auth_url(
  scopes = load_scopes(),
  redirect_uri = NULL,
  team_code = NULL,
  state = NULL,
  client_id = NULL
)
}
\arguments{
\item{scopes}{character, scopes to request.}

\item{redirect_uri}{character, the uri to which the user should be directed
after authorization. If this is NULL (default), the user will be directed
to the default redirection set up in the app.}

\item{team_code}{character, a team code to restrict the user to (in case they
have multiple Slack teams authorized in their browser).}

\item{state}{character, a code to send to your redirect_uri indicating a
state. It is recommended to use a non-human-readable format for this
string.}

\item{client_id}{character, the client_id of a Slack app. If this is not
provided, the function will use the built-in R4DS Slack app.}
}
\value{
character, an authorization URL.
}
\description{
Generate an authorization URL to add a team manually. Useful,
for example, for authorizing a team in a Shiny app.
}
\examples{
auth_url()
auth_url(redirect_uri = "http://127.0.0.1:4242")
auth_url(redirect_uri = "http://127.0.0.1:4242", team_code = "T6UC1DKJQ")
auth_url(state = "aabbCCddeeFF")
}
\concept{management}
