% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_app.R
\name{create_custom_app}
\alias{create_custom_app}
\title{Custom App Helper}
\usage{
create_custom_app(
  app_name,
  redirect_urls = NULL,
  description = NULL,
  long_description = NULL
)
}
\arguments{
\item{app_name}{character; a name to give to your custom app.}

\item{redirect_urls}{character; urls to which the app can be redirected
during authentication. We recommend including localhost on a specific port
(eg, http://127.0.0.1:4242) for testing, and often a shinyapps.io url.}

\item{description}{character; an optional one-sentence description of your
app. If this parameter is NULL, an adequate description is provided.}

\item{long_description}{character; an optional paragraph (or more) giving
more details about your app. This field supports markdown. If this
parameter is NULL, an adequate long description is provided.}
}
\value{
The app manifest as an editable R object, invisibly.
}
\description{
Create a manifest for a custom Slack App. See the Slack Apps vignette for
details about using custom Slack Apps.
}
\examples{
create_custom_app(
  app_name = "my_cool_app",
  redirect_urls = c("http://127.0.0.1:4242", "myapp.shinyapps.io")
)
}
\concept{management}
