% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_app.R
\name{custom_app_yaml}
\alias{custom_app_yaml}
\title{Prepare an App Manifest}
\usage{
custom_app_yaml(app_manifest)
}
\arguments{
\item{app_manifest}{list; an app manifest such as the one returned by
\code{\link{create_custom_app}}.}
}
\value{
The manifest in yaml format (invisibly).
}
\description{
Convert an app manifest list object to a yaml for upload.
}
\examples{
my_manifest <- create_custom_app("myApp")
my_manifest$`_metadata`$minor_version <- 1L
custom_app_yaml(my_manifest)
}
