% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_conversations_list}
\alias{get_conversations_list}
\title{Get Channel List}
\usage{
get_conversations_list(
  type = c("public_channel", "private_channel", "mpim", "im"),
  ...,
  token = Sys.getenv("SLACK_API_TOKEN"),
  max_results = Inf,
  max_calls = Inf
)
}
\arguments{
\item{type}{character, the types of channels to include in the fetch.}

\item{...}{additional parameters passed on to
\code{\link[slackcalls]{post_slack}}}

\item{token}{character, api token issued by slack}

\item{max_results}{integer, the maximum number of results to return (total).}

\item{max_calls}{integer, the maximum number of separate API calls to make while constructing the response.}
}
\value{
tibble
}
\description{
Fetch channel information from the Slack API.
}
\concept{get}
