% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_managment.R
\name{load_team_dcf}
\alias{load_team_dcf}
\title{Load Team(s) From File}
\usage{
load_team_dcf(team = "user", file = "~/.slackr", verbose = TRUE)
}
\arguments{
\item{team}{character, name of team, Default: 'user'}

\item{file}{character, Path to dcf containing slackr compatible fields, Default: '~/.slackr'}

\item{verbose}{logical, Print message after loading, Default: TRUE}
}
\description{
Load teams from local file that contains slack API definitions
that are slackr compatible
}
\concept{files}
