% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_managment.R
\name{load_teams}
\alias{load_teams}
\title{Load Team(s) From File}
\usage{
load_teams(file = "~/.slackteams", verbose = TRUE)
}
\arguments{
\item{file}{character, Path to json containing tokens, Default: '~/.slackteams'}

\item{verbose}{logical, Print message after loading, Default: TRUE}
}
\description{
Load teams from local file that contains slack tokens
}
\concept{files}
