% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scope_managment.R
\name{scopes}
\alias{scopes}
\title{Scope Management}
\usage{
scopes(
  file = system.file("scopes.yml", package = "slackteams"),
  which = "slackverse"
)
}
\arguments{
\item{file}{character; a path to a yaml file with scopes defined.}

\item{which}{character; the group of scopes to use within that yaml.}
}
\value{
A character vector of scopes.
}
\description{
Determine scopes from a file.
}
