% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{slackteams}
\alias{slackteams}
\title{Manage slackteams File}
\usage{
slackteams(file, verbose = TRUE, auto_unbox = TRUE, pretty = TRUE, ...)
}
\arguments{
\item{file}{character, path to file}

\item{verbose}{logical, print to console, Default: TRUE}

\item{auto_unbox}{logical, automatically unbox all atomic vectors of length 1, Default: TRUE}

\item{pretty}{logical adds indentation whitespace to JSON output, Default: TRUE}

\item{...}{arguments to pass to \link[jsonlite:read_json]{write_json}}
}
\description{
Upload or Update json file containing slack tokens
}
\concept{files}
