% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_managment.R
\name{upload_team_file}
\alias{upload_team_file}
\alias{update_team_file}
\title{Manage Team Files}
\usage{
upload_team_file(file)

update_team_file(file, verbose = TRUE, auto_unbox = TRUE,
  pretty = TRUE, ...)
}
\arguments{
\item{file}{character, path to file}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}

\item{auto_unbox}{PARAM_DESCRIPTION, Default: TRUE}

\item{pretty}{PARAM_DESCRIPTION, Default: TRUE}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Upload or Update json file containing slackr-app keys
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
