% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_channel}
\alias{validate_channel}
\title{Validate Channel Label}
\usage{
validate_channel(channel)
}
\arguments{
\item{channel}{character, channel label}
}
\value{
character
}
\description{
Validate channel label and convert it to Slack channel ID
}
