% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_managment.R
\name{add_team_interactive}
\alias{add_team_interactive}
\title{Interactive Team Managment}
\usage{
add_team_interactive(scopes = load_scopes(), verbose = TRUE)
}
\arguments{
\item{scopes}{character, scopes to request. Must include "users:read",
"channels:read", "groups:read", "im:read", and "mpim:read" at minimum.}

\item{verbose}{logical, Print messages to console, Default: TRUE}
}
\description{
Add a team interactively
}
\details{
Launch a browser window to interactively grant slackteams permission
to act on your behalf on a Slack team.
}
\note{
This function does not currently work in an Rstudio Server setup. We
are exploring options to remedy this situation.
}
