% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_methods.R
\name{get_teams}
\alias{get_teams}
\alias{get_team_env}
\alias{get_active_team}
\alias{get_team_creds}
\alias{get_team_users}
\alias{get_team_channels}
\title{Query Team Information}
\usage{
get_teams()

get_team_env()

get_active_team()

get_team_creds(team)

get_team_users(team)

get_team_channels(team)
}
\arguments{
\item{team}{character, name of team}
}
\value{
character
}
\description{
Functions to extract information about teams
}
\concept{get}
