% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_methods.R
\name{get_teams}
\alias{get_teams}
\alias{get_team_env}
\alias{get_active_team}
\alias{get_team_creds}
\alias{get_team_users}
\alias{get_team_channels}
\alias{get_channel_members}
\alias{get_channel_info}
\alias{get_member_name}
\alias{get_activeteam_id}
\title{Query Team Information}
\usage{
get_teams()

get_team_env()

get_active_team()

get_team_creds(team = get_active_team())

get_team_users(team, fields = c("id", "name", "real_name"))

get_team_channels(team, fields = NULL)

get_channel_members(channel)

get_channel_info(channel)

get_member_name(id)

get_activeteam_id()
}
\arguments{
\item{team}{character, name of team}

\item{fields}{character, columns to return in get_team_channels and get_team_users, NULL will return all the columns}

\item{channel}{character, name of a channel}

\item{id}{character, unique member id}
}
\value{
character
}
\description{
Functions to extract information about teams
}
\details{
Channels can be a public channel, a private channel a direct message or group direct message. To get the names of the channels
the user has access to use \link[=get_team_channels]{get_team_channels(fields = 'name')}.
}
\concept{get}
