% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_managment.R
\name{activate_team}
\alias{activate_team}
\alias{add_team}
\alias{remove_team}
\title{Team Managment Functions}
\usage{
activate_team(team, verbose = TRUE)

add_team(team, memberid, key)

remove_team(team)
}
\arguments{
\item{team}{character, team name}

\item{verbose}{logical, Print messages to console, Default: TRUE}

\item{memberid}{chracter, member id of user in the team}

\item{key}{character, slackr-app key unique to the memberid}
}
\description{
Manage teams that can be accessed
}
\concept{managment}
