% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_management.R
\name{activate_team}
\alias{activate_team}
\alias{add_team}
\alias{add_team_token}
\alias{remove_team}
\title{Team Management Functions}
\usage{
activate_team(team, verbose = TRUE)

add_team(team, token)

add_team_token(team, token, verbose = TRUE)

remove_team(team)
}
\arguments{
\item{team}{character, team name}

\item{verbose}{logical, Print messages to console, Default: TRUE}
}
\description{
Manage teams that can be accessed
}
\concept{management}
