% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_managment.R
\name{upload_team_file}
\alias{upload_team_file}
\alias{update_team_file}
\title{Manage Team Files}
\usage{
upload_team_file(file)

update_team_file(file, verbose = TRUE, auto_unbox = TRUE, pretty = TRUE, ...)
}
\arguments{
\item{file}{character, path to file}

\item{verbose}{logical, print to console, Default: TRUE}

\item{auto_unbox}{logical, automatically unbox all atomic vectors of length 1, Default: TRUE}

\item{pretty}{logical adds indentation whitespace to JSON output, Default: TRUE}

\item{...}{arguments to pass to \link[jsonlite:read_json]{write_json}}
}
\description{
Upload or Update json file containing slackr-app keys
}
\concept{files}
