% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_channel}
\alias{validate_channel}
\title{Validate Channel Label}
\usage{
validate_channel(channel)
}
\arguments{
\item{channel}{character, channel label}
}
\value{
character
}
\description{
Validate that channel label exists in active team and
convert it to the corresponding slack channel ID
}
\concept{management}
